Add files via upload
To promote transparency and reproducibility of our experiments, we made the dataset consisting of the headlines comprising 26,270 headlines from January 1st, 2012, to April 1st, 2021 available. We performed sentiment analysis on the headlines with the following sentiment classifiers: FinBERT, CrudeBERT, and CrudeBERT+. In addition, the proprietary "Event Sentiment Score" of RavenPack has also been included as an additional benchmark for evaluation purposes.